Reorganize a little files
